apps/matching-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
motor==3.3.2
scikit-learn==1.3.2
numpy==1.26.3
python-dotenv==1.0.0
httpx==0.26.0
